+++ lib/jobmanager.py
@@ -4,7 +4,8 @@ import datetime
 import logging
 import os
 import datetime
-from typing import Any, Callable
+import time
+from typing import Any, Callable, Optional, Union
 
 import qpack
 from enodo.jobs import *
@@ -32,8 +33,14 @@ class EnodoJob:
     __slots__ = ('rid', 'series_name', 'job_config',
                  'job_data', 'send_at', 'error', 'worker_id')
 
-    def __init__(self, rid, series_name, job_config, job_data=None,
-                 send_at=None, error=None, worker_id=None):
+    def __init__(self,
+                 rid: Union[int, str],
+                 series_name: str,
+                 job_config: SeriesJobConfigModel,
+                 job_data: Optional[dict] = None,
+                 send_at: Optional[int] = None,
+                 error: Optional[str] = None,
+                 worker_id: Optional[str] = None):
         if not isinstance(
                 job_data, EnodoJobDataModel) and job_data is not None:
             raise Exception('Unknown job data value')
@@ -51,8 +58,6 @@ class EnodoJob:
         for slot in cls.__slots__:
             if isinstance(getattr(job, slot), datetime.datetime):
                 resp[slot] = int(getattr(job, slot).timestamp())
-            elif isinstance(getattr(job, slot), SeriesJobConfigModel):
-                resp[slot] = getattr(job, slot)
             else:
                 resp[slot] = getattr(job, slot)
         return resp
@@ -206,7 +211,7 @@ class EnodoJobManager:
             if cls._update_queue_cb is not None:
                 await cls._update_queue_cb(
                     SUBSCRIPTION_CHANGE_TYPE_DELETE, job.rid)
-        job.send_at = datetime.datetime.now()
+        job.send_at = time.time()
         job.worker_id = worker_id
         cls._active_jobs.append(job)
         cls._active_jobs_index[job.rid] = job
@@ -477,7 +482,7 @@ class EnodoJobManager:
             job_data = EnodoJobRequestDataModel(
                 job_id=job.rid, job_config=job.job_config,
                 series_name=job.series_name,
-                series_config=series.series_config,
+                series_config=series.config,
                 series_state=series.state,
                 siridb_ts_units=ServerState.siridb_ts_unit)
             data = qpack.packb(job_data.serialize())